New View_Relaunch method
The method View_Relaunch is intended for applications that handle
relaunch operation themselves. The relaunch operation happens when an
application is already running and are requested to launch with a
parameter (i.e. requesting a browser to open a specific URI).

Open-webOS-DCO-1.0-Signed-off-by: Jose Dapena Paz <jose.dapena@palm.com>

Change-Id: I22447f4200dc55b9e34fe468d3012475424e6651