Provide a pkgconfig configuration file
Having a pkgconfig configuration files makes it a lot easier to detect wether the package
and it's header files are installed and to provide correct include paths rather than
adding them statically to every related component.

NOTE: this is changing the cmake project name to match the name of the built library.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>